Moved global parameters into a separate configuration file.
Instead of having path to images, the back image and the valid image
extensions I moved them into dodo.conf which, more or less, matches python
code. NOTE! The values (everything after =) are actually evaluated using
eval! This must be changed at some point. It's fine now, as this is a rough
draft/prototype but, in general, using eval on raw user input is a big NO NO